# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "TG Summarizer", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/ai/models": {
      "get": {
        "tags": ["ai"],
        "summary": "Api List Models",
        "operationId": "ai-api_list_models",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Ai-Api List Models"
                }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/ai/summary": {
      "post": {
        "tags": ["ai"],
        "summary": "Api Summary",
        "operationId": "ai-api_summary",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SummaryRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Ai-Api Summary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/ai/summary/prompt": {
      "post": {
        "tags": ["ai"],
        "summary": "Api Summary Prompt",
        "operationId": "ai-api_summary_prompt",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SummaryRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Ai-Api Summary Prompt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/ai/summary/stream": {
      "post": {
        "tags": ["ai"],
        "summary": "Api Summary Stream",
        "operationId": "ai-api_summary_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SummaryRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/ai/chat/stream": {
      "post": {
        "tags": ["ai"],
        "summary": "Api Chat Stream",
        "operationId": "ai-api_chat_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChatRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/ai/embeddings": {
      "post": {
        "tags": ["ai"],
        "summary": "Api Embeddings",
        "operationId": "ai-api_embeddings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EmbedRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Ai-Api Embeddings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/ai/translate": {
      "post": {
        "tags": ["ai"],
        "summary": "Api Translate",
        "operationId": "ai-api_translate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TranslateRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Ai-Api Translate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatMessage": {
        "properties": {
          "role": { "type": "string", "title": "Role" },
          "text": { "type": "string", "title": "Text" }
        },
        "type": "object",
        "required": ["role", "text"],
        "title": "ChatMessage"
      },
      "ChatRequest": {
        "properties": {
          "channels": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Channels",
            "default": []
          },
          "postsText": {
            "type": "string",
            "title": "Poststext",
            "default": ""
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "English"
          },
          "model": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Model"
          },
          "temperature": {
            "type": "number",
            "title": "Temperature",
            "default": 0.7
          },
          "message": { "type": "string", "title": "Message" },
          "history": {
            "items": { "$ref": "#/components/schemas/ChatMessage" },
            "type": "array",
            "title": "History",
            "default": []
          },
          "ragMode": {
            "type": "boolean",
            "title": "Ragmode",
            "default": false
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "gemini"
          }
        },
        "type": "object",
        "required": ["message"],
        "title": "ChatRequest"
      },
      "EmbedRequest": {
        "properties": {
          "texts": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Texts"
          },
          "model": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Model"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "gemini"
          }
        },
        "type": "object",
        "required": ["texts"],
        "title": "EmbedRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SummaryRequest": {
        "properties": {
          "channels": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Channels"
          },
          "postsText": { "type": "string", "title": "Poststext" },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "English"
          },
          "model": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Model"
          },
          "temperature": {
            "type": "number",
            "title": "Temperature",
            "default": 0.7
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "gemini"
          }
        },
        "type": "object",
        "required": ["channels", "postsText"],
        "title": "SummaryRequest"
      },
...
      "TranslateRequest": {
        "properties": {
          "posts": {
            "items": {
              "additionalProperties": { "type": "string" },
              "type": "object"
            },
            "type": "array",
            "title": "Posts"
          },
          "targetLanguage": { "type": "string", "title": "Targetlanguage" },
          "model": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Model"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "gemini"
          }
        },
        "type": "object",
        "required": ["posts", "targetLanguage"],
        "title": "TranslateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/api/v1/login/access-token" }
        }
      }
    }
  }
}